requirements.txt
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
primary==0.0.1.dev1
psycopg2==2.9.3
pycountry==22.3.5
requests==2.27.1
sqlparse==0.4.2
twilio==7.9.1
urllib3==1.26.9
pytz==2022.1
